fixed 1 typo and 1 bug in get_good_EEG leading to 0% good.  Switched to ecg_epochs_around_feedback 2 -> original
There were cases where 0% ECG was occurring in the output incorrectly.  This was due to multiple factors incliding two bugs/typos in get_goodECG where Ngood1[i] was computed.  I don't think the ecg_epochs_around_feedback2 was getting all the data either (not sure why the date split implementation failed), so I reverted to the original function, which will take longer, but gets the correct data.  Curious if this fixes the issues.